Notebooks/5_genai_calorie_calculator/requirements.txt
google-generativeai
python-dotenv
streamlit
langchain
langchain_google_genai
Pillow